+++ crml/transaction-payment/Cargo.toml
@@ -3,6 +3,10 @@ name = "crml-transaction-payment"
 version = "1.0.0"
 authors = ["Centrality Developers <support@centrality.ai>"]
 edition = "2018"
+license = "GPL-3.0"
+homepage = "https://cennznetdocs.com"
+repository = "https://github.com/cennznet/cennznet"
+description = "CENNZnet pallet to manage transaction payments"
 
 [dependencies]
 codec = { version = "1.2.0", package = "parity-scale-codec", default-features = false, features = ["derive"] }
